stars falling too fast fixed
Stars were falling too fast. This is because after they touched a player or the Janitor, they would reset back to fall down again. This was because the stars maintained their acceleration downwards. Which meant they got faster and faster as time went on. Now if a star hits a player or the janitor, their velocity gets reset to 0, so that doesn't happen. Code is a bit unclean, fix that if time.